+++ Filter/Filter.go
@@ -1,4 +1,4 @@
-package Utils
+package Filter
 
 import "fmt"
